failure-backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
motor
dnspython
requests
python-multipart
yake
